Remove some duplicate code, rename legacy.ts to legacyAsyncStorag.ts
